Moved static pages into component-staticPages folder
