+++ app/routes.js
@@ -11,9 +11,9 @@ import CookieNotice from './components/global/CookieNotice'
 import LandingRedirect from './components/pages/LandingRedirect'
 import LoginPage from './components/pages/Login'
 import LogoutPage from './components/pages/Logout/index'
-import AuthorGuide from './components/pages/StaticPages/AuthorGuide'
-import ReviewerGuide from './components/pages/StaticPages/ReviewerGuide'
-import ContactUs from './components/pages/StaticPages/ContactUs'
+import AuthorGuide from '../packages/component-staticPages/client/pages/AuthorGuide'
+import ReviewerGuide from '../packages/component-staticPages/client/pages/ReviewerGuide'
+import ContactUs from '../packages/component-staticPages/client/pages/ContactUs'
 
 import SubmissionWizard from './components/pages/SubmissionWizard'
 import ErrorPage from './components/pages/Error'
