+++ packages/component-staticPages/client/components/SideNav.js
