+++ packages/component-staticPages/client/pages/AuthorGuide/EditorialProcess.md
