+++ packages/component-staticPages/client/pages/AuthorGuide/Fees.js
