+++ packages/component-staticPages/client/pages/AuthorGuide/Fees.md
