+++ packages/component-staticPages/client/pages/AuthorGuide/Full.md
