+++ packages/component-staticPages/client/pages/AuthorGuide/JournalCharts.data.js
@@ -90,11 +90,11 @@ export const chart4 = [
   ['Q1 2017', 0, 24, 31, 41, 45],
   ['Q2 2017', 0, 24, 32, 41, 50],
   ['Q3 2017', 0, 26, 34, 46, 50],
-  ['Q4 2017', 0, 28, 36, 47, 50], 
+  ['Q4 2017', 0, 28, 36, 47, 50],
   /* 2018 */
   ['Q1 2018', 0, 25, 34, 43, 50],
   ['Q2 2018', 0, 27, 36, 48, 50],
-  ['Q3 2018', 0, 28, 37, 47, 50], 
+  ['Q3 2018', 0, 28, 37, 47, 50],
   ['Q4 2018', 0, 25, 32, 42, 50],
 ]
