+++ packages/component-staticPages/client/pages/AuthorGuide/JournalCharts.js
