+++ packages/component-staticPages/client/pages/AuthorGuide/JournalMetrics.md
