+++ packages/component-staticPages/client/pages/AuthorGuide/JournalPolicies.md
