+++ packages/component-staticPages/client/pages/AuthorGuide/Post.md
