+++ packages/component-staticPages/client/pages/AuthorGuide/RequiredInfoTable.js
