+++ packages/component-staticPages/client/pages/AuthorGuide/Revised.md
