+++ packages/component-staticPages/client/pages/AuthorGuide/Types.js
