+++ packages/component-staticPages/client/pages/AuthorGuide/Types.md
