+++ packages/component-staticPages/client/pages/AuthorGuide/index.js
