+++ packages/component-staticPages/client/pages/ReviewerGuide/ReviewingPolicies.md
