+++ packages/component-staticPages/client/pages/index.js
@@ -8,8 +8,8 @@ import { th } from '@pubsweet/ui-toolkit'
 import ErrorBoundary from 'global/ErrorBoundary'
 import media from 'global/layout/media'
 import NavigationDropdown from 'ui/atoms/NavigationDropdown'
-import SideNav from './SideNav'
-import ScrollToTop from './ScrollToTop'
+import SideNav from '../components/SideNav'
+import ScrollToTop from '../components/ScrollToTop'
 
 const TopNavContainer = styled(Box).attrs({ mx: -3 })`
   border-bottom: ${th('borderWidth')} ${th('borderStyle')} ${th('colorBorder')};
