+++ packages/component-staticPages/client/pages/index.md
