+++ packages/component-staticPages/client/pages/index.test.js
